feat: Add separate process functions for squat and bicep curl rep counting with enhanced visibility checks and feedback
